Introducing cron script
The current cron script has been set to run every miinute to reduce the time taken to test the effects each change makes.
To ensure successful working of the cron job there is an assoicated .bash file as well as updates to the wallpaper.py file.
The readme has been updated with very minor instructions on how to configure the cron job and a .gitignore file has been included too.
* 20 Jan 2019 - maintaining anonymised cron.d/desktopchanger_cron
* 20 Jan 2019 - anonymising desktopchanger.bash file