Argument parsing improvements
Shifted the location of the argument parsing portion down and updated it so that specifying an image is an optional argument.
Changed the WallpaperChanger class to store the currently displayed wallpaper (old_wallpaper).
Encapsulated the bash commands in a try block.
Improved the main script body for readibility and minor changes due to changes in how the class works.