Project Restructure
The project restructure has been to two root scripts which  can be run. One for updating the desktop effects (wallpaper, in future panels) and the other for loading a file with image paths (in future these images are analysed first). Other files have been added which will also be filled out.
Additoinally:
* Readme has been partially updated
* A license has been applied and,
* the .gitignore file has been updated
* an empty requirements.txt that needs atleast the pyyaml project

* 20 Jan 2019 - Maintaining anonymised desktopchanger_cron file
* 20 Jan 2019 - Anonymising config.yaml file