CSV file introduction
The CSV file of images has been introduced with a new script to search a folder and create the list.
The desktopchanger.py module can now read the contents of the .csv file and choose from this list.
The config.yaml file is now being read in to find this wallpapers folder and processed from there.

* 20 Jan 2019 - Deleting unnecessary reloadimages.py file