Sunrise Equation
The Sunrise Equation found on Wikipedia was used to approximately calculate the sunrise and sunset times.
This is stored as a separate module with updates to the desktopchanger module to allow the equation to be run.
Additionally this data is saved to file where it can be read from instead of calculating this result each run.
The utils module includes a module for yaml IO given that there is now multiple yaml files which can be used.
New configuration is needed to run this program namely the latitude and longitude values for the current position are required.